build_bkey {Vizumap}	R Documentation

Build a bivariate colour key

Description

This function links data and a colour palette. Numerical bounds are added to the 3 x 3 colour grid.

Usage

build_bkey(data, palette = "BlueYellow", terciles = FALSE, flipAxis = FALSE)

Arguments

data
A data frame.

palette
Name of colour palette or character vector of hex colour codes created with build_palette function. Colour palette names include BlueYellow, CyanMagenta, BlueRed and GreenBlue.

terciles
A logical value. This provides the option to define numerical bounds for the colour key grid using terciles instead of equal intervals.

flipAxis
A logical value. Whether to place the axis on the opposite sides or not.

Details

A key for the bivariate map is not automatically generated with build_bmap and must be made using build_bkey. It is important that the arguments passed to build_bkey match those passed to build_bmap. The map and key can be viewed together using attach_key.

See Also

attach_key

Examples

data(us_data)
data(us_geo)
poverty <- read.uv(data = us_data, estimate = "pov_rate", error = "pov_moe")

#use a prepared palette and terciles
key <- build_bkey(data = poverty, terciles = TRUE)
view(key)

#use a created palette
p <- build_palette(name = "usr", colrange =
 list(colour = c("darkblue", "chartreuse4"), difC = c(3, 4)))
key <- build_bkey(data = poverty, palette = p)
view(key)

[Package Vizumap version 1.2.0 Index]