build_palette {Vizumap}	R Documentation

Build a palette

Description

This function prepares one of the four included colour palettes or builds a new colour palette.

Usage

build_palette(
  name,
  colrange = list(colour = NULL, difC = NULL),
  flipVertical = FALSE,
  flipHorizontal = FALSE
)

Arguments

name
Name of colour palette or usr for option to design a new palette. Colour palette names include BlueYellow, CyanMagenta, BlueRed and GreenBlue.

colrange
List with a character vector of length two and a numeric vector of length two.

flipVertical
Whether the palette should be flipped vertically (ie. replace top portion with bottom portion)

flipHorizontal
Whether the palette should be flipped horizontally (ie. replace left portion with right portion)

colour
A character vector of two colour names from the colours() range or valid hexadecimal colors.

difC
A numeric vector of two integers 1, 2, 3 or 4. Values control how much a colour changes in value across the grid. One corresponds with a small change in colour value, and four corresponds with a large change in colour value.

Details

Note that colrange only needs to be specified if name = "usr." When choosing colours, it is best to avoid light colours or tints as these will lead to a colour palette lacking noticeable differences across the 3 x 3 colour grid.

Examples

#use one of four prepared colour palettes
p <- build_palette(name = "CyanMagenta")
view(p)

#design a new palette
p <- build_palette(name = "usr", colrange =
 list(colour = c("darkblue", "chartreuse4"), difC = c(3, 4)))
view(p)

[Package Vizumap version 1.2.0 Index]