exc_9596 {Vizumap}	R Documentation

exc_9596

Description

A dataset that contains estimates of Total Suspended Sediment (TSS), the error, i.e. standard deviation (TSS_error), the exceedance probabilities that correspond to thresholds of 837 mg/L and 2204 mg/L, respectively for each polygon or reach contributing area (RCA). These thresholds were determined from Bartley et al. (2012). The TSS estimates, error and exceedance probabilities were derived from the model developed and described in Gladish et al. (2016).

Usage

exc_9596

Format

A data frame comprising 411 rows (one for each RCA) and 5 columns.

References

Bartley, R., Speirs, W. J., Ellis, T. W., and Waters, D. K. (2012). A review of sediment and nutrient concentration data from Australia for use in catchment water quality models. Marine Pollution Bulletin 65, 101–116. doi:10.1016/J.MARPOLBUL.2011.08.009 Gladish, D. W., Kuhnert, P. M., Pagendam, D. E., Wikle, C. K., Bartley, R., Searle, R. D., Ellis, R. J., Dougall, C., Turner, R. D. R., Lewis, S. E., Bainbridge, Z. T., and Brodie, J. E. (2016). Spatio-temporal assimilation of modelled catchment loads with monitoring data in the Great Barrier Reef. The Annals of Applied Statistics 10, 1590–1618. doi:10.1214/16- AOAS950

[Package Vizumap version 1.2.0 Index]