findNbounds

Find the bins for the colour grid

findNbounds(
  data,
  estimate,
  error,
  terciles,
  expertR_est = NA,
  expertR_err = NA
)

Arguments

data

Dataset from which bounds are to be determined.

estimate

Name of the estimate column.

error

Name of the error column.

terciles

Should terciles be calculated? (Default: FALSE).

expertR_est

A vector consisting of the range of expert values for the estimate (Default: NA).

expertR_err

A vector consisting of the range of expert values for the error (Default: NA).

Details

Expert ranges for the estimate and error can be supplied to assist with the comparison of multiple figures. By default these are set to NA, which will result in the bounds being directly computed from the data. Note, if the expert values do not span the range of the data, this function will default to finding the bounds of the data.