pdist {Vizumap}	R Documentation

pdist

Description

Calculates the probability of exceedance for a given distribution, threshold and distribution parameters.

Usage

pdist(pname, th, args)

Arguments

pname
Distribution to be used which is embedded through the quote command. This can comprise an existing function from the stats package e.g. pnorm. Alternatively this can be user defined.

th
Threshold for calculating the \[Pr(X > th)\]

args
Arguments that correspond to the specified distribution (but excluding the threshold.

Examples

pname <- quote({ pnorm(q, mean, sd, lower.tail = FALSE) })
pdist(pname = pname, th = 3, args = list(mean = 2, sd = 1))

[Package Vizumap version 1.2.0 Index]