pixelate {Vizumap}	R Documentation

Pixelate polygons

Description

This function divides a shapefile into pixels so it can be used to create a pixel map with build_pmap.

Usage

pixelate(shapefile = NULL, file = NULL, layer = NULL, pixelSize = 2, id = "id")

Arguments

shapefile
An object of class "SpatialPolygons" or "SpatialPolygonsDataFrame".

file
A shapefile pathway.

layer
Name of shapefile layer (see documentation for readOGR for more information).

pixelSize
An integer 1, 2 or 3. One corresponds to the smallest pixel size, and three corresponds to the largest.

id
A name which will be given to the new ID column. This ID corresponds to the slot ID in the shapefile.

Details

This function can take several minutes to run depending on the size of the shapefile.

Examples

data(us_geo)
ca_geo <- subset(us_geo, us_geo@data$STATE == "06")
pix <- pixelate(ca_geo, id = "region")

[Package Vizumap version 1.2.0 Index]