us_data {Vizumap}	R Documentation

USA Poverty Data

Description

A dataset containing estimated poverty rates among American families for 2015. Included is a margin of error for each estimate, which can be used to calculate the estimate's 90% confidence interval.

Usage

us_data

Format

A data frame with 3142 elements and 5 variables:

GEO_ID

14-digit code that identifies summary level of data and region

GEO.id2

FIPS code

GEO.display.label

geographic name for geographic area

pov_rate

percentage of families whose income was below the poverty level

pov_moe

margin of error for poverty estimate

Source

http://factfinder.census.gov/

[Package Vizumap version 1.2.0 Index]