us_geo {Vizumap}	R Documentation

USA shapefile

Description

A dataset that produces a USA county map. It can be used to visualise data in the us_data data frame spatially.

Usage

us_geo

Format

A large spatial polygons data frame with 3142 elements and 7 variables:

GEO_ID

14-digit code that identifies summary level of data and region

STATE

state FIPS code

COUNTY

county FIPS code

NAME

region name

LSAD

legal statistical area description

SHAPE_AREA

area measurement in square meters

SHAPE_LEN

perimeter measurement in meters

Source

http://factfinder.census.gov/

[Package Vizumap version 1.2.0 Index]