view {Vizumap}	R Documentation

View a Vizumap object

Description

This function views maps, keys, palettes and animations created with Vizumap.

Usage

view(obj)

Arguments

obj
An object from build_bmap, build_gmap, build_pmap, build_bkey, build_gkey, build_palette, or animate.

Details

For maps, keys and palettes, assigning the output from view to a variable name saves the output as an object of class "ggplot". Animations created with animate are viewed through an internet browser as html objects.

Examples

gb <- build_palette(name = "GreenBlue")
view(gb)

#ggplot object
p <- view(gb)

[Package Vizumap version 1.2.0 Index]