Pixelate polygons

This function divides a shapefile into pixels so it can be used to create a pixel map with build_pmap.

pixelate(geoData = NULL, file = NULL, layer = NULL, pixelSize = 2, id = "id")

Arguments

geoData
An object of class "SpatialPolygons" or "SpatialPolygonsDataFrame".

file
A shapefile pathway.

layer
Name of geoData layer (see documentation for readOGR for more information).

pixelSize
An integer 1, 2 or 3. One corresponds to the smallest pixel size, and three corresponds to the largest.

id
A name which will be given to the new ID column. This ID corresponds to the slot ID in the spatial data.

Details

This function can take several minutes to run depending on the size of the shapefile. Within this function, the projection of the spatial object is removed and then returned to the original projection.

Examples

data(us_geo)
ca_geo <- subset(us_geo, us_geo@data$STATE == "06")
pix <- pixelate(ca_geo, id = "region")